refactor : StatisticsResult 의 matchTicket method 간소화.
StatisticsResult의 matchTickets 메서드를 stream으로 변경. bonusBallMatch를 enum의 맴버로 추가.
